requirements.txt
# FattureInCloud MCP Server - Python Dependencies
# SDK ufficiale Fatture in Cloud (verificato 2026-01-14)
fattureincloud-python-sdk==2.1.3

# MCP SDK Ufficiale (Anthropic)
mcp

# Environment variables
python-dotenv==1.0.1

# HTTP/SSE server dependencies (for remote deployment)
uvicorn>=0.27.0
starlette>=0.36.0

# HTTP client (usato da SDK)
urllib3>=1.25.3

# Date handling
python-dateutil>=2.8.2
